update delete api

+++ app/api/tasks/[id]/route.ts
@@ -1,13 +1,20 @@
 import { auth } from "@/auth";
 
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
+
     const session = await auth();
+    if (!session) {
+        return new Response('You are not authenticated', { status: 401 });
+    }
+
     const deletedId = params.id;
-    const response = await fetch(process.env.TASKS_BASE_API + '/' + deletedId, {
+    const url = process.env.TASKS_BASE_API + '/' + deletedId;
+    const response = await fetch(url, {
         headers: {
             'Authorization': 'Bearer ' + session?.user?.accessToken,
         },
         method: 'DELETE'
     });
+
     return new Response(null, { status: response.status });
 }
